requirements.txt
Flask
flask-cors
psycopg2-binary
gunicorn
tensorflow
numpy
matplotlib
pillow
scikit-learn
opencv-python
requests